+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/commands.py
@@ -23,6 +23,7 @@ from ._util import (
     get_sql_firewall_rules_operations,
     get_sql_replication_links_operations,
     get_sql_restorable_dropped_databases_operations,
+    get_sql_server_connection_policies_operations,
     get_sql_server_keys_operations,
     get_sql_servers_operations,
     get_sql_server_usages_operations,
@@ -299,4 +300,15 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
             c.command('show', 'get')
             c.command('list', 'list_by_server')
             c.command('delete', 'delete')
+            c.generic_update_command('update', 'get', 'create_or_update')
+
+    server_connection_policies_operations = create_service_adapter(
+        'azure.mgmt.sql.operations.server_connection_policies_operations',
+        'ServerConnectionPoliciesOperations'
+    )
+
+    with ServiceGroup(__name__, get_sql_server_connection_policies_operations, server_connection_policies_operations,
+                      custom_path) as s:
+        with s.group('sql server conn-policy') as c:
+            c.command('show', 'get')
             c.generic_update_command('update', 'get', 'create_or_update')