+++ src/command_modules/azure-cli-sql/HISTORY.rst
@@ -4,8 +4,9 @@ Release History
 ===============
 
 2.0.16
-++++++++++
+++++++
 * Added az sql db list-usages and az sql db show-usage commands.
+* Added sql server conn-policy show/update commands.
 
 2.0.15
 ++++++